# src/main/resources/openapi.yaml
openapi: 3.0.1
info:
  title: "Payment API"
  version: "1.0"

paths:
  /payments:
    post:
      summary: "Create payment"
      operationId: "createPayment"
      description: "The transaction ID must be unique, otherwise a concurrent processing exception will be thrown. The amount must be less than or equal to 1000, because the simplified balance check always returns 1000 as the available amount."
      requestBody:
        description: "Data of the payment"
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentCreateRequest"
      responses:
        "201":
          description: "Payment created successfully"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentCreateResponse"
        "400":
          description: "Bad request"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: "Server error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    PaymentCreateRequest:
      type: object
      required:
        - clientId
        - transactionId
        - amount
      properties:
        clientId:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        transactionId:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        amount:
          type: integer
          format: int64
          example: 1

    PaymentCreateResponse:
      type: object
      properties:
        transactionId:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"

    ErrorResponse:
      type: object
      properties:
        errorCode:
          type: integer
          example: 400
        message:
          type: string
          example: "Bad request: missing property"